+++ rbxmk/cmd_upload-asset.go
@@ -45,6 +45,7 @@ func (c *UploadAssetCommand) Run(opt snek.Options) error {
 	world, err := InitWorld(WorldOpt{
 		WorldFlags:       WorldFlags{Debug: false},
 		ExcludeRoots:     true,
+		ExcludeEnums:     true,
 		ExcludeLibraries: true,
 		ExcludeProgram:   true,
 	})